pyproject.toml
[project]
name = "ctofbrain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.1",
    "fastapi>=0.124.0",
    "langchain>=1.1.3",
    "langchain-anthropic>=1.2.0",
    "langchain-chroma>=1.0.0",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.2",
    "langchain-groq>=1.1.0",
    "langchain-huggingface>=1.1.0",
    "langgraph>=1.0.4",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pydicom>=3.0.1",
    "pypdf>=6.4.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.52.1",
    "tensorflow>=2.20.0",
    "tf-keras>=2.20.1",
    "transformers>=4.57.3",
    "uvicorn>=0.38.0",
]
